Register validations using @validation function decorator
instead of hand-enumerating them in separate data structures.

Introduce 'validation' function decorator accepting (policies, predicate)
constructor arguments, along with rx() function to construct a predicate
matching a particular regular expression. The rx() function is to eliminate
the need to distinguish at runtime between a callable predicate and a regular
expression: now the relevant policy collections hold only predicates.

Move the two rule-selector functions above the validation functions so we can
pass them to the validation decorator.

Decorate each (referenced) policy check function with @validation decorator
naming the policy collection in which it should be registered, along with the
predicate to use for that function. The policy collection is taken from the
common_policies, proprietary_policies or opensource_policies tuple in which
the policy check function was originally hand-registered; the predicate is
either the predicate function that was originally registered, or
rx(expression) where 'expression' is the originally-registered regular
expression.

In checker.check_file(), remove the case handling 'pat' as a regular
expression, leaving only the case of a callable predicate.