Add per-policy-function decoder support.
The default is utf8(), which means that most policy functions will be
passed str data as they expect. But a given policy function may specify raw()
or utf16() instead if desired.

Introduce a cache into checker.check_file() so that running multiple policy
functions with the same decoder over the same file data will only have to run
that decoder once.

Read commit-message files in text mode ("r") because we don't check them for
line endings.

In Python 3, the binary() predicate function needs to search for '\0' or
b'\0', depending on whether it's passed a bytes object or a str object.