Try harder to decode raw bytes for files we think should be text.
Add a maybe_text() decoder for use with windows_line_endings() and
unix_line_endings(). One of those is called for *every* file, even binary
files. maybe_text() bugs out if it thinks the raw byte stream is binary, else
it tries to decode as UTF-8, else it uses chardet to try to detect the actual
encoding.

Test for binary() here in the decoder, rather than in either of those policy
functions.

Use decoder=raw for valid_llsd() since llsd.parse() expects bytes rather than
str.

Add **kwds argument to each of the decoder functions. If the relevant decoder
raises UnicodeDecodeError, call it again with errors='replace' so we have
something to pass to the policy function.