Use pre-commit project to install hooks
This changeset is a large departure from the existing layout of the
git-hooks repository. Testing has been added for all hooks, and each
unit of behavior is now broken out into a separate small script.

The `install` file has been removed as it has been replaced with an open
source project, `pre-commit`, which takes care of dependency and hook
installation. `pre-commit` also provides a more flexible way for
consumers of this repository to select which hooks they want to use,
which may be useful for third-party developers that only want to use
some of our hooks.